refactor: 레포지토리를 인터페이스와 JDBC 구현체로 분리
